requirements.txt
PlexApi
neo4j
requests
python-dotenv